# docs/server/swagger.yaml
components:
  schemas:
...
    client.Client:
      properties:
        name:
          $ref: '#/components/schemas/client.MCPClient'
      type: object
    client.MCPClient:
      type: string
      x-enum-varnames:
      - RooCode
      - Cline
      - Cursor
      - VSCodeInsider
      - VSCode
      - ClaudeCode
      - ClaudeDesktop
...
    permissions.NetworkPermissions:
      description: Network defines network permissions
      properties:
        outbound:
          $ref: '#/components/schemas/permissions.OutboundNetworkPermissions'
      type: object
    permissions.OutboundNetworkPermissions:
      description: Outbound defines outbound network permissions
      properties:
        allow_host:
          description: AllowHost is a list of allowed hosts
          items:
            type: string
          type: array
          uniqueItems: false
        allow_port:
          description: AllowPort is a list of allowed ports
          items:
            type: integer
          type: array
          uniqueItems: false
        insecure_allow_all:
          description: InsecureAllowAll allows all outbound network connections
          type: boolean
      type: object
    permissions.Profile:
      description: PermissionProfile is the permission profile to use
      properties:
        name:
          description: Name is the name of the profile
          type: string
        network:
          $ref: '#/components/schemas/permissions.NetworkPermissions'
        read:
          description: |-
            Read is a list of mount declarations that the container can read from
            These can be in the following formats:
            - A single path: The same path will be mounted from host to container
            - host-path:container-path: Different paths for host and container
            - resource-uri:container-path: Mount a resource identified by URI to a container path
          items:
            type: string
          type: array
          uniqueItems: false
        write:
          description: |-
            Write is a list of mount declarations that the container can write to
            These follow the same format as Read mounts but with write permissions
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
    registry.EnvVar:
      properties:
        default:
          description: |-
            Default is the value to use if the environment variable is not explicitly provided
            Only used for non-required variables
          type: string
        description:
          description: Description is a human-readable explanation of the variable's
            purpose
          type: string
        name:
          description: Name is the environment variable name (e.g., API_KEY)
          type: string
        required:
          description: |-
            Required indicates whether this environment variable must be provided
            If true and not provided via command line or secrets, the user will be prompted for a value
          type: boolean
        secret:
          description: |-
            Secret indicates whether this environment variable contains sensitive information
            If true, the value will be stored as a secret rather than as a plain environment variable
          type: boolean
      type: object
    registry.ImageMetadata:
      description: Server details
      properties:
        args:
          description: |-
            Args are the default command-line arguments to pass to the MCP server container.
            These arguments will be prepended to any command-line arguments provided by the user.
          items:
            type: string
          type: array
          uniqueItems: false
        custom_metadata:
          additionalProperties: {}
          description: CustomMetadata allows for additional user-defined metadata
          type: object
        description:
          description: Description is a human-readable description of the server's
            purpose and functionality
          type: string
        docker_tags:
          description: DockerTags lists the available Docker tags for this server
            image
          items:
            type: string
          type: array
          uniqueItems: false
        env_vars:
          description: EnvVars defines environment variables that can be passed to
            the server
          items:
            $ref: '#/components/schemas/registry.EnvVar'
          type: array
          uniqueItems: false
        image:
          description: Image is the Docker image reference for the MCP server
          type: string
        metadata:
          $ref: '#/components/schemas/registry.Metadata'
        name:
          description: |-
            Name is the identifier for the MCP server, used when referencing the server in commands
            If not provided, it will be auto-generated from the image name
          type: string
        permissions:
          $ref: '#/components/schemas/permissions.Profile'
        provenance:
          $ref: '#/components/schemas/registry.Provenance'
        repository_url:
          description: RepositoryURL is the URL to the source code repository for
            the server
          type: string
        status:
          description: The Status indicates whether the server is currently active
            or deprecated
          type: string
        tags:
          description: Tags are categorization labels for the server to aid in discovery
            and filtering
          items:
            type: string
          type: array
          uniqueItems: false
        target_port:
          description: TargetPort is the port for the container to expose (only applicable
            to SSE and Streamable HTTP transports)
          type: integer
        tier:
          description: Tier represents the tier classification level of the server,
            e.g., "official" or "community" driven
          type: string
        tools:
          description: Tools is a list of tool names provided by this MCP server
          items:
            type: string
          type: array
          uniqueItems: false
        transport:
          description: Transport defines the communication protocol for the server
            (stdio, sse, or streamable-http)
          type: string
      type: object
    registry.Metadata:
      description: Metadata contains additional information about the server such
        as popularity metrics
      properties:
        last_updated:
          description: LastUpdated is the timestamp when the server was last updated,
            in RFC3339 format
          type: string
        pulls:
          description: Pulls indicates how many times the server image has been downloaded
          type: integer
        stars:
          description: Stars represents the popularity rating or number of stars for
            the server
          type: integer
      type: object
    registry.Provenance:
      description: Provenance contains verification and signing metadata
      properties:
        attestation:
          $ref: '#/components/schemas/registry.VerifiedAttestation'
        cert_issuer:
          type: string
        repository_ref:
          type: string
        repository_uri:
          type: string
        runner_environment:
          type: string
        signer_identity:
          type: string
        sigstore_url:
          type: string
      type: object
    registry.Registry:
      description: Full registry data
      properties:
        last_updated:
          description: LastUpdated is the timestamp when the registry was last updated,
            in RFC3339 format
          type: string
        servers:
          additionalProperties:
            $ref: '#/components/schemas/registry.ImageMetadata'
          description: Servers is a map of server names to their corresponding server
            definitions
          type: object
        version:
          description: Version is the schema version of the registry
          type: string
      type: object
    registry.VerifiedAttestation:
      properties:
        predicate: {}
        predicate_type:
          type: string
      type: object
...
    v1.bulkClientRequest:
      properties:
        names:
          description: Names is the list of client names to operate on.
          items:
            type: string
            x-enum-varnames:
            - RooCode
            - Cline
            - Cursor
            - VSCodeInsider
            - VSCode
            - ClaudeCode
            - ClaudeDesktop
          type: array
          uniqueItems: false
      type: object
...
    v1.createClientRequest:
      properties:
        name:
          description: Name is the type of the client to register.
          type: string
          x-enum-varnames:
          - RooCode
          - Cline
          - Cursor
          - VSCodeInsider
          - VSCode
          - ClaudeCode
          - ClaudeDesktop
      type: object
    v1.createClientResponse:
      properties:
        name:
          description: Name is the type of the client that was registered.
          type: string
          x-enum-varnames:
          - RooCode
          - Cline
          - Cursor
          - VSCodeInsider
          - VSCode
          - ClaudeCode
          - ClaudeDesktop
      type: object
...
    v1.getRegistryResponse:
      description: Response containing registry details
      properties:
        last_updated:
          description: Last updated timestamp
          type: string
        name:
          description: Name of the registry
          type: string
        registry:
          $ref: '#/components/schemas/registry.Registry'
        server_count:
          description: Number of servers in the registry
          type: integer
        version:
          description: Version of the registry schema
          type: string
      type: object
...
    v1.getServerResponse:
      description: Response containing server details
      properties:
        server:
          $ref: '#/components/schemas/registry.ImageMetadata'
      type: object
...
    v1.listServersResponse:
      description: Response containing a list of servers
      properties:
        servers:
          description: List of servers in the registry
          items:
            $ref: '#/components/schemas/registry.ImageMetadata'
          type: array
          uniqueItems: false
      type: object
...
    v1.registryInfo:
      description: Basic information about a registry
      properties:
        last_updated:
          description: Last updated timestamp
          type: string
        name:
          description: Name of the registry
          type: string
        server_count:
          description: Number of servers in the registry
          type: integer
        version:
          description: Version of the registry schema
          type: string
      type: object
    v1.registryListResponse:
      description: Response containing a list of registries
      properties:
        registries:
          description: List of registries
          items:
            $ref: '#/components/schemas/v1.registryInfo'
          type: array
          uniqueItems: false
      type: object
...
info:
  description: This is the ToolHive API server.
  title: ToolHive API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /api/v1beta/clients:
    get:
      description: List all registered clients in ToolHive
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/client.Client'
                type: array
          description: OK
      summary: List all clients
      tags:
      - clients
    post:
      description: Register a new client with ToolHive
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/v1.createClientRequest'
        description: Client to register
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.createClientResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                type: string
          description: Invalid request
      summary: Register a new client
      tags:
      - clients
  /api/v1beta/clients/{name}:
    delete:
      description: Unregister a client from ToolHive
      parameters:
      - description: Client name to unregister
        in: path
        name: name
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                type: string
          description: Invalid request
      summary: Unregister a client
      tags:
      - clients
  /api/v1beta/clients/register:
    post:
      description: Register multiple clients with ToolHive
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/v1.bulkClientRequest'
        description: Clients to register
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/v1.createClientResponse'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                type: string
          description: Invalid request
      summary: Register multiple clients
      tags:
      - clients
  /api/v1beta/clients/unregister:
    post:
      description: Unregister multiple clients from ToolHive
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/v1.bulkClientRequest'
        description: Clients to unregister
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                type: string
          description: Invalid request
      summary: Unregister multiple clients
      tags:
      - clients
...
  /api/v1beta/registry:
    get:
      description: Get a list of the current registries
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.registryListResponse'
          description: OK
      summary: List registries
      tags:
      - registry
    post:
      description: Add a new registry
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "501":
          content:
            application/json:
              schema:
                type: string
          description: Not Implemented
      summary: Add a registry
      tags:
      - registry
  /api/v1beta/registry/{name}:
    delete:
      description: Remove a specific registry
      parameters:
      - description: Registry name
        in: path
        name: name
        required: true
        schema:
          type: string
      responses:
        "204":
          content:
            application/json:
              schema:
                type: string
          description: No Content
        "404":
          content:
            application/json:
              schema:
                type: string
          description: Not Found
      summary: Remove a registry
      tags:
      - registry
    get:
      description: Get details of a specific registry
      parameters:
      - description: Registry name
        in: path
        name: name
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.getRegistryResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                type: string
          description: Not Found
      summary: Get a registry
      tags:
      - registry
  /api/v1beta/registry/{name}/servers:
    get:
      description: Get a list of servers in a specific registry
      parameters:
      - description: Registry name
        in: path
        name: name
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.listServersResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                type: string
          description: Not Found
      summary: List servers in a registry
      tags:
      - registry
  /api/v1beta/registry/{name}/servers/{serverName}:
    get:
      description: Get details of a specific server in a registry
      parameters:
      - description: Registry name
        in: path
        name: name
        required: true
        schema:
          type: string
      - description: ImageMetadata name
        in: path
        name: serverName
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.getServerResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                type: string
          description: Not Found
      summary: Get a server from a registry
      tags:
      - registry
...